6-size.c

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,11 @@ int main(void)
 	long int z;
 	long long int m;
 	float n;
-printf("size of a char: %lu byte(s)\n", (unsigned long)sizeof(x));
-printf("size of a int: %lu byte(s)\n", (unsigned long)sizeof(y));
-printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(z));
-printf("size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(m));
-printf("size of a float: %lu byte(s)\n", (unsigned long)sizeof(n));
-return (0);
+
+	printf("size of a char: %d byte(s)\n", sizeof(x));
+	printf("size of a int: %d byte(s)\n", sizeof(y));
+	printf("size of a long int: %d byte(s)\n", sizeof(z));
+	printf("size of a long long int: %d byte(s)\n", sizeof(m));
+	printf("size of a float: %d byte(s)\n", sizeof(n);
+	return (0);
 }
